Modify workload_profile structure
